refactor(Event) : 결과 타입 long에서 int로 통일

+++ src/main/java/christmas/model/event/ChristmasEvent.java
@@ -10,7 +10,7 @@ public class ChristmasEvent implements Event{
     private static final String CHRISTMAS_EVENT = "크리스마스 디데이 할인";
     @Override
     public DiscountResult calculateDiscountAmount(Order order) {
-        long discountAmount = 0;
+        int discountAmount = 0;
 
         LocalDate startTime = LocalDate.of(2023, 12, 1);
         LocalDate endTime = LocalDate.of(2023, 12, 25);
