+++ src/main/java/christmas/model/event/SpecialDayEvent.java
@@ -10,7 +10,7 @@ public class SpecialDayEvent implements Event {
 
     @Override
     public DiscountResult calculateDiscountAmount(Order order) {
-        long discountAmount = 0;
+        int discountAmount = 0;
         LocalDate startTime = LocalDate.of(2023, 12, 1);
         LocalDate endTime = LocalDate.of(2023, 12, 25);
         List<Integer> days = SpecialDay.getDays();
